REFACTOR : Rank 패키지 이동
