+++ src/main/java/kr/hhplus/be/server/interfaces/rank/api/RankController.java
@@ -1,9 +1,12 @@
-package kr.hhplus.be.server.interfaces.rank;
+package kr.hhplus.be.server.interfaces.rank.api;
 
-import kr.hhplus.be.server.application.rank.RankCriteria;
-import kr.hhplus.be.server.application.rank.RankFacade;
-import kr.hhplus.be.server.application.rank.RankResult;
+import java.time.LocalDate;
+
+import kr.hhplus.be.server.domain.rank.RankCommand;
+import kr.hhplus.be.server.domain.rank.RankInfo;
+import kr.hhplus.be.server.domain.rank.RankService;
 import kr.hhplus.be.server.interfaces.ApiResponse;
+import kr.hhplus.be.server.interfaces.rank.RankResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -12,11 +15,11 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class RankController {
 
-    private final RankFacade rankFacade;
+    private final RankService rankService;
 
     @GetMapping("/api/v1/products/ranks")
     public ApiResponse<RankResponse.PopularProducts> getPopularProducts() {
-        RankResult.PopularProducts popularProducts = rankFacade.getPopularProducts(RankCriteria.PopularProducts.ofTop5Days3());
+        RankInfo.PopularProducts popularProducts = rankService.cachedPopularProducts(RankCommand.PopularProducts.ofTop5Days3(LocalDate.now()));
         return ApiResponse.success(RankResponse.PopularProducts.of(popularProducts));
     }
 }
